Copy derictory name and correct update server list

+++ src/main/java/ru/prokhorov/model/FileDelete.java
@@ -4,21 +4,25 @@ import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
-import java.io.Serializable;
 import java.nio.file.Files;
-import java.nio.file.Path;
 import java.nio.file.Paths;
 
 @Data
 @Slf4j
 public class FileDelete implements  AbstractMessage {
 
+
     private String deletePach;
 
     public FileDelete(String fileName) throws IOException {
         deletePach = "serverFiles";
         String deleteP = deletePach + "/" + fileName;
-        Files.delete(Paths.get(deleteP));
+        boolean existsFile = Files.exists(Paths.get(deleteP));
+        if(existsFile) {
+            Files.delete(Paths.get(deleteP));
+        }else{
+            log.debug("File exists!");
+        }
     }
 
     @Override
